refactor(auth): move CLI pairing routes
