+++ docs/context/architecture/composition.md
@@ -19,7 +19,8 @@ related:
 # Application composition
 
 `bootstrap.create_app(role)` is the FastAPI composition root. `api.py` hosts the ordered route table;
-the auth, Catalog, web, and admin modules define concern-specific `APIRouter` blocks that `api.py`
+the auth module defines separate email OTP and CLI pairing blocks, while the Catalog, web, and admin
+modules define their concern-specific `APIRouter` blocks that `api.py`
 appends at their legacy registration points. It then calls the factory once at EOF so the deployed
 and documented `treg.api:app` import path remains the default `all` role.
